Relying on module outputs for metrics mappings
