+++ aws/cloudwatch/metrics/mappings.go
@@ -1,0 +1,73 @@
+package metrics
+
+import (
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/cloudwatch/types"
+	"github.com/nullstone-io/deployment-sdk/block"
+	"k8s.io/utils/strings/slices"
+)
+
+type MappingGroups []MappingGroup
+
+type MappingGroup struct {
+	Name     string                   `json:"name"`
+	Type     block.MetricDatasetType  `json:"type"`
+	Unit     string                   `json:"unit"`
+	Mappings map[string]MetricMapping `json:"mappings"`
+}
+
+type MetricMapping struct {
+	AccountId  string                  `json:"account_id"`
+	Stat       string                  `json:"stat"`
+	Namespace  string                  `json:"namespace"`
+	MetricName string                  `json:"metric_name"`
+	Dimensions MetricMappingDimensions `json:"dimensions"`
+}
+
+type MetricMappingDimensions map[string]string
+
+func (d MetricMappingDimensions) ToAws() []types.Dimension {
+	dims := make([]types.Dimension, 0)
+	for k, v := range d {
+		dims = append(dims, types.Dimension{
+			Name:  aws.String(k),
+			Value: aws.String(v),
+		})
+	}
+	return dims
+}
+
+func (g MappingGroups) BuildMetricQueries(metrics []string, periodSec int32) []types.MetricDataQuery {
+	queries := make([]types.MetricDataQuery, 0)
+	for _, grp := range g {
+		if len(metrics) < 1 || slices.Contains(metrics, grp.Name) {
+			// This Metric Group was specified in the list of requested metrics
+			// Let's build a query and add it
+			for id, mapping := range grp.Mappings {
+				queries = append(queries, types.MetricDataQuery{
+					Id:        aws.String(id),
+					AccountId: aws.String(mapping.AccountId),
+					MetricStat: &types.MetricStat{
+						Period: aws.Int32(periodSec),
+						Stat:   aws.String(mapping.Stat),
+						Metric: &types.Metric{
+							Namespace:  aws.String(mapping.Namespace),
+							MetricName: aws.String(mapping.MetricName),
+							Dimensions: mapping.Dimensions.ToAws(),
+						},
+					},
+				})
+			}
+		}
+	}
+	return queries
+}
+
+func (g MappingGroups) FindGroupByMetricId(id string) *MappingGroup {
+	for _, grp := range g {
+		if _, ok := grp.Mappings[id]; ok {
+			return &grp
+		}
+	}
+	return nil
+}
