--- aws/ecs/metrics_mappings.go
@@ -1,65 +1,0 @@
-package ecs
-
-import (
-	"github.com/nullstone-io/deployment-sdk/aws/cloudwatch"
-	"github.com/nullstone-io/deployment-sdk/block"
-)
-
-var (
-	MetricMappings = cloudwatch.MetricMappingGroups{
-		{
-			Name: "cpu",
-			Type: block.MetricDatasetTypeUsage,
-			Unit: "vCPU",
-			Mappings: map[string]cloudwatch.MetricMapping{
-				"cpu_reserved": {
-					Stat:      "Average",
-					Namespace: "ECS/ContainerInsights",
-					Name:      "CpuReserved",
-				},
-				"cpu_average": {
-					Stat:      "Average",
-					Namespace: "ECS/ContainerInsights",
-					Name:      "CpuUtilized",
-				},
-				"cpu_min": {
-					Stat:      "Minimum",
-					Namespace: "ECS/ContainerInsights",
-					Name:      "CpuUtilized",
-				},
-				"cpu_max": {
-					Stat:      "Maximum",
-					Namespace: "ECS/ContainerInsights",
-					Name:      "CpuUtilized",
-				},
-			},
-		},
-		{
-			Name: "memory",
-			Type: block.MetricDatasetTypeUsage,
-			Unit: "MiB",
-			Mappings: map[string]cloudwatch.MetricMapping{
-				"cpu_reserved": {
-					Stat:      "Average",
-					Namespace: "ECS/ContainerInsights",
-					Name:      "CpuReserved",
-				},
-				"cpu_average": {
-					Stat:      "Average",
-					Namespace: "ECS/ContainerInsights",
-					Name:      "CpuUtilized",
-				},
-				"cpu_min": {
-					Stat:      "Minimum",
-					Namespace: "ECS/ContainerInsights",
-					Name:      "CpuUtilized",
-				},
-				"cpu_max": {
-					Stat:      "Maximum",
-					Namespace: "ECS/ContainerInsights",
-					Name:      "CpuUtilized",
-				},
-			},
-		},
-	}
-)
